even moar refactor
