--- agent/agent_linux.go
@@ -1,90 +1,0 @@
-/*
-Copyright 2022 AmidaWare LLC.
-
-Licensed under the Tactical RMM License Version 1.0 (the “License”).
-You may only use the Licensed Software in accordance with the License.
-A copy of the License is available at:
-
-https://license.tacticalrmm.com
-
-*/
-
-package agent
-
-import (
-	rmm "github.com/amidaware/rmmagent/shared"
-	"github.com/kardianos/service"
-	trmm "github.com/wh1te909/trmm-shared"
-)
-
-
-
-// windows only below TODO add into stub file
-
-
-
-
-
-
-
-
-
-func (a *Agent) GetPython(force bool) {}
-
-type SchedTask struct{ Name string }
-
-func (a *Agent) PatchMgmnt(enable bool) error { return nil }
-
-func (a *Agent) CreateSchedTask(st SchedTask) (bool, error) { return false, nil }
-
-func DeleteSchedTask(name string) error { return nil }
-
-func ListSchedTasks() []string { return []string{} }
-
-func (a *Agent) GetEventLog(logName string, searchLastDays int) []rmm.EventLogMsg {
-	return []rmm.EventLogMsg{}
-}
-
-func (a *Agent) GetServiceDetail(name string) trmm.WindowsService { return trmm.WindowsService{} }
-
-func (a *Agent) ControlService(name, action string) rmm.WinSvcResp {
-	return rmm.WinSvcResp{Success: false, ErrorMsg: "/na"}
-}
-
-func (a *Agent) EditService(name, startupType string) rmm.WinSvcResp {
-	return rmm.WinSvcResp{Success: false, ErrorMsg: "/na"}
-}
-
-func (a *Agent) GetInstalledSoftware() []trmm.WinSoftwareList { return []trmm.WinSoftwareList{} }
-
-func (a *Agent) ChecksRunning() bool { return false }
-
-func (a *Agent) RunTask(id int) error { return nil }
-
-func (a *Agent) InstallChoco() {}
-
-func (a *Agent) InstallWithChoco(name string) (string, error) { return "", nil }
-
-func (a *Agent) GetWinUpdates() {}
-
-func (a *Agent) InstallUpdates(guids []string) {}
-
-func (a *Agent) installMesh(meshbin, exe, proxy string) (string, error) {
-	return "not implemented", nil
-}
-
-func CMDShell(shell string, cmdArgs []string, command string, timeout int, detached bool) (output [2]string, e error) {
-	return [2]string{"", ""}, nil
-}
-
-func CMD(exe string, args []string, timeout int, detached bool) (output [2]string, e error) {
-	return [2]string{"", ""}, nil
-}
-
-func (a *Agent) GetServices() []trmm.WindowsService { return []trmm.WindowsService{} }
-
-func (a *Agent) Start(_ service.Service) error { return nil }
-
-func (a *Agent) Stop(_ service.Service) error { return nil }
-
-func (a *Agent) InstallService() error { return nil }
