+++ agent/patching/patching_linux.go
@@ -1,0 +1,7 @@
+package patching
+
+func PatchMgmnt(enable bool) error { return nil }
+
+func GetWinUpdates() {}
+
+func InstallUpdates(guids []string) {}
