--- agent/software/software.go
@@ -1,3 +1,0 @@
-package software
-
-func SendSoftware() {}
